requirements.txt
# Fuzzy matching and string comparison
python-Levenshtein>=0.21.1,<0.22.0
fuzzywuzzy>=0.18.0

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
python-multipart>=0.0.6
pydantic>=2.4.2
requests>=2.31.0
httpx>=0.25.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiohttp>=3.9.0
python-slugify>=8.0.0

# Square integration
squareup==39.0.0.20241120

# Web scraping
beautifulsoup4==4.12.2
PyYAML==6.0.1
cachetools==4.2.4
ratelimit==2.2.1

# AWS integration
boto3==1.34.34
botocore==1.34.34
urllib3<1.27,>=1.25.4

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.10.0
flake8>=6.1.0
isort>=5.12.0